+++ tests/lib/astarte_device_sdk/integration/device_caching/src/device_caching_test.c
@@ -149,6 +149,30 @@ static bool astarte_individual_is_equal(astarte_individual_t first, astarte_indi
 ZTEST_SUITE(astarte_device_sdk_device_caching, NULL, device_caching_test_setup,
     device_caching_test_before, device_caching_test_after, device_caching_test_teardown); // NOLINT
 
+ZTEST_F(astarte_device_sdk_device_caching, test_device_caching_synchronization) // NOLINT
+{
+    bool sync = false;
+    astarte_result_t ares = ASTARTE_RESULT_OK;
+
+    ares = astarte_device_caching_synchronization_get(&sync);
+    zassert_equal(ares, ASTARTE_RESULT_NOT_FOUND, "Res:%s", astarte_result_to_name(ares));
+    zassert_equal(sync, false, "sync variable has been modified");
+
+    sync = true;
+    ares = astarte_device_caching_synchronization_get(&sync);
+    zassert_equal(ares, ASTARTE_RESULT_NOT_FOUND, "Res:%s", astarte_result_to_name(ares));
+    zassert_equal(sync, true, "sync variable has been modified");
+
+    sync = true;
+    ares = astarte_device_caching_synchronization_set(sync);
+    zassert_equal(ares, ASTARTE_RESULT_OK, "Res:%s", astarte_result_to_name(ares));
+
+    sync = false;
+    ares = astarte_device_caching_synchronization_get(&sync);
+    zassert_equal(ares, ASTARTE_RESULT_OK, "Res:%s", astarte_result_to_name(ares));
+    zassert_equal(sync, true, "Sync variable not set correctly");
+}
+
 ZTEST_F(astarte_device_sdk_device_caching, test_device_caching_store_introspection) // NOLINT
 {
     astarte_result_t ares = ASTARTE_RESULT_OK;
@@ -248,7 +272,7 @@ ZTEST_F(astarte_device_sdk_device_caching, test_device_caching_store_load_proper
     zassert_equal(ares, ASTARTE_RESULT_OK, "Res:%s", astarte_result_to_name(ares));
 
     read_major = 0;
-    read_individual = (astarte_individual_t){ 0 };
+    read_individual = (astarte_individual_t) { 0 };
     ares = astarte_device_caching_property_load(
         property_2.interface_name, property_2.path, &read_major, &read_individual);
     zassert_equal(ares, ASTARTE_RESULT_OK, "Res:%s", astarte_result_to_name(ares));
@@ -258,7 +282,7 @@ ZTEST_F(astarte_device_sdk_device_caching, test_device_caching_store_load_proper
     astarte_device_caching_property_destroy_loaded(read_individual);
 
     read_major = 0;
-    read_individual = (astarte_individual_t){ 0 };
+    read_individual = (astarte_individual_t) { 0 };
     ares = astarte_device_caching_property_load(
         property_3.interface_name, property_3.path, &read_major, &read_individual);
     zassert_equal(ares, ASTARTE_RESULT_OK, "Res:%s", astarte_result_to_name(ares));
@@ -269,7 +293,7 @@ ZTEST_F(astarte_device_sdk_device_caching, test_device_caching_store_load_proper
 
     // The first property has been overwritten by the last one
     read_major = 0;
-    read_individual = (astarte_individual_t){ 0 };
+    read_individual = (astarte_individual_t) { 0 };
     ares = astarte_device_caching_property_load(
         property_4.interface_name, property_4.path, &read_major, &read_individual);
     zassert_equal(ares, ASTARTE_RESULT_OK, "Res:%s", astarte_result_to_name(ares));
@@ -319,7 +343,7 @@ ZTEST_F(astarte_device_sdk_device_caching, test_device_caching_iterate) // NOLIN
     zassert_equal(ares, ASTARTE_RESULT_OK, "Res:%s", astarte_result_to_name(ares));
 
     astarte_device_caching_property_iter_t iter = { 0 };
-    ares = astarte_device_caching_property_iterator_init(&iter);
+    ares = astarte_device_caching_property_iterator_new(&iter);
     zassert_equal(ares, ASTARTE_RESULT_OK, "Res:%s", astarte_result_to_name(ares));
 
     interface_name_size = 0U;
@@ -388,6 +412,8 @@ ZTEST_F(astarte_device_sdk_device_caching, test_device_caching_iterate) // NOLIN
 
     ares = astarte_device_caching_property_iterator_next(&iter);
     zassert_equal(ares, ASTARTE_RESULT_NOT_FOUND, "Res:%s", astarte_result_to_name(ares));
+
+    astarte_device_caching_property_iterator_destroy(iter);
 }
 
 ZTEST_F(astarte_device_sdk_device_caching, test_device_caching_iterate_empty) // NOLINT
@@ -395,8 +421,9 @@ ZTEST_F(astarte_device_sdk_device_caching, test_device_caching_iterate_empty) //
     astarte_result_t ares = ASTARTE_RESULT_OK;
 
     astarte_device_caching_property_iter_t iter = { 0 };
-    ares = astarte_device_caching_property_iterator_init(&iter);
+    ares = astarte_device_caching_property_iterator_new(&iter);
     zassert_equal(ares, ASTARTE_RESULT_NOT_FOUND, "Res:%s", astarte_result_to_name(ares));
+    astarte_device_caching_property_iterator_destroy(iter);
 }
 
 ZTEST_F(astarte_device_sdk_device_caching, test_device_caching_delete) // NOLINT
@@ -472,7 +499,7 @@ ZTEST_F(astarte_device_sdk_device_caching, test_device_caching_delete) // NOLINT
 
     // Loop over all the stored properties
     astarte_device_caching_property_iter_t iter = { 0 };
-    ares = astarte_device_caching_property_iterator_init(&iter);
+    ares = astarte_device_caching_property_iterator_new(&iter);
     zassert_equal(ares, ASTARTE_RESULT_OK, "Res:%s", astarte_result_to_name(ares));
 
     interface_name_size = ARRAY_SIZE(interface_name_buffer);
@@ -540,10 +567,11 @@ ZTEST_F(astarte_device_sdk_device_caching, test_device_caching_delete) // NOLINT
 
     ares = astarte_device_caching_property_iterator_next(&iter);
     zassert_equal(ares, ASTARTE_RESULT_NOT_FOUND, "Res:%s", astarte_result_to_name(ares));
+    astarte_device_caching_property_iterator_destroy(iter);
 
     // Loop over all the stored properties
-    iter = (astarte_device_caching_property_iter_t){ 0 };
-    ares = astarte_device_caching_property_iterator_init(&iter);
+    iter = (astarte_device_caching_property_iter_t) { 0 };
+    ares = astarte_device_caching_property_iterator_new(&iter);
     zassert_equal(ares, ASTARTE_RESULT_OK, "Res:%s", astarte_result_to_name(ares));
 
     interface_name_size = ARRAY_SIZE(interface_name_buffer);
@@ -590,6 +618,7 @@ ZTEST_F(astarte_device_sdk_device_caching, test_device_caching_delete) // NOLINT
 
     ares = astarte_device_caching_property_iterator_next(&iter);
     zassert_equal(ares, ASTARTE_RESULT_NOT_FOUND, "Res:%s", astarte_result_to_name(ares));
+    astarte_device_caching_property_iterator_destroy(iter);
 }
 
 ZTEST_F(astarte_device_sdk_device_caching, test_device_caching_get_properties_string) // NOLINT